requirements.txt
streamlit
pymongo[srv]
passlib[bcrypt]
python-dotenv
pandas
bcrypt==3.2.0
plotly